refactor(api): extract OAuth state PKCE seal helpers
Move AES-GCM seal/open into oauth_state_pkce.go with focused tests
to keep oauth.go smaller and PKCE crypto easier to review.
